feat: publish `gear-eth-bridge-primitives` to crates.io (#4734)
