+++ pallets/gear-eth-bridge/primitives/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "pallet-gear-eth-bridge-primitives"
+description = "Primitives for Gear Ethereum Bridge Pallet"
+documentation = "https://docs.rs/pallet-gear-eth-bridge-primitives"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+# Operational deps.
+parity-scale-codec.workspace = true
+scale-info.workspace = true
+serde = { workspace = true, features = ["derive"], optional = true }
+
+# Gear deps.
+gprimitives.workspace = true
+
+# Substrate deps.
+binary-merkle-tree.workspace = true
+
+[features]
+default = ["std"]
+std = [
+    # Operational deps.
+    "parity-scale-codec/std",
+    "scale-info/std",
+    "dep:serde",
+
+    # Gear deps.
+    "gprimitives/serde",
+
+    # Substrate deps.
+    "binary-merkle-tree/std",
+]
